Add ex 5.5-5.11: -Blog post form is displayed only when appropriate -Separate blog post form into its own component -Hide/show details button for each blog -Fix name not being shown for new blog posts -Like button functionality -Sort blog posts by number of likes -Delete blog post functionality

+++ part4/blog/controllers/blogs.js
@@ -19,9 +19,9 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if(!decodedToken.id){
-    return response.status(401).json({error: 'token invalid'})
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
   }
 
   const user = await User.findById(decodedToken.id);
@@ -39,6 +39,11 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
@@ -46,13 +51,18 @@ blogsRouter.delete("/:id", async (request, response) => {
 blogsRouter.put("/:id", async (request, response, next) => {
   const body = request.body;
 
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
   const blog = {
     likes: body.likes,
   };
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
     new: true,
-  });
+  }).populate("user", { username: 1, name: 1 });
   response.json(updatedBlog);
 });
